requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==2.2
django-cors-headers==3.1.1
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.1.0
idna==2.9
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
requests==2.23.0
shortuuid==1.0.1
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
